requirements.txt
appnope==0.1.3
asttokens==2.0.8
backcall==0.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
decorator==5.1.1
docker-pycreds==0.4.0
et-xmlfile==1.1.0
executing==0.10.0
filelock==3.8.0
gitdb==4.0.9
GitPython==3.1.27
huggingface-hub==0.9.1
idna==3.3
ipython==8.4.0
jedi==0.18.1
matplotlib-inline==0.1.3
numpy==1.23.2
openai==0.23.0
openpyxl==3.0.10
packaging==21.3
pandas==1.4.3
pandas-stubs==1.4.3.220815
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
promise==2.3
prompt-toolkit==3.0.30
protobuf==3.20.1
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
PyYAML==6.0
regex==2022.8.17
requests==2.28.1
sentry-sdk==1.9.5
setproctitle==1.3.2
shortuuid==1.0.9
six==1.16.0
smmap==5.0.0
stack-data==0.4.0
tokenizers==0.12.1
tqdm==4.64.0
traitlets==5.3.0
transformers==4.21.2
types-pytz==2022.1.2
typing_extensions==4.3.0
urllib3==1.26.11
wandb==0.13.1
wcwidth==0.2.5
